get rid of spark names
